backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.77.1",
    "fastapi>=0.128.1",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "uvicorn>=0.40.0",
]
